LPS-74032 rename dyslexic package
