+++ modules/apps/foundation/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/document/contributor/AssetEntryDocumentContributor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.internal.contributor.document;
+package com.liferay.portal.search.internal.document.contributor;
 
 import com.liferay.asset.kernel.AssetRendererFactoryRegistryUtil;
 import com.liferay.asset.kernel.model.AssetEntry;
