+++ modules/apps/foundation/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/document/contributor/AssetTagDocumentContributor.java
@@ -12,15 +12,19 @@
  * details.
  */
 
-package com.liferay.portal.search.internal.contributor.document;
+package com.liferay.portal.search.internal.document.contributor;
 
+import com.liferay.asset.kernel.model.AssetTag;
+import com.liferay.asset.kernel.service.AssetTagLocalService;
 import com.liferay.portal.kernel.model.BaseModel;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.GroupedModel;
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.DocumentContributor;
 import com.liferay.portal.kernel.search.Field;
-import com.liferay.portal.kernel.service.GroupLocalService;
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.ListUtil;
+import com.liferay.portal.kernel.util.Portal;
+
+import java.util.List;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -29,42 +33,34 @@ import org.osgi.service.component.annotations.Reference;
  * @author Michael C. Han
  */
 @Component(immediate = true, service = DocumentContributor.class)
-public class GroupedModelDocumentContributor implements DocumentContributor {
+public class AssetTagDocumentContributor implements DocumentContributor {
 
 	@Override
 	public void contribute(Document document, BaseModel baseModel) {
-		if (!(baseModel instanceof GroupedModel)) {
-			return;
-		}
+		String className = document.get(Field.ENTRY_CLASS_NAME);
 
-		GroupedModel groupedModel = (GroupedModel)baseModel;
+		long classPK = GetterUtil.getLong(document.get(Field.ENTRY_CLASS_PK));
 
-		document.addKeyword(
-			Field.GROUP_ID, getSiteGroupId(groupedModel.getGroupId()));
-		document.addKeyword(Field.SCOPE_GROUP_ID, groupedModel.getGroupId());
-	}
+		long classNameId = portal.getClassNameId(className);
 
-	protected Group getSiteGroup(long groupId) {
-		Group group = groupLocalService.fetchGroup(groupId);
+		List<AssetTag> assetTags = assetTagLocalService.getTags(
+			classNameId, classPK);
 
-		if ((group != null) && group.isLayout()) {
-			group = group.getParentGroup();
-		}
+		String[] assetTagNames = ListUtil.toArray(
+			assetTags, AssetTag.NAME_ACCESSOR);
 
-		return group;
-	}
+		document.addText(Field.ASSET_TAG_NAMES, assetTagNames);
 
-	protected long getSiteGroupId(long groupId) {
-		Group group = getSiteGroup(groupId);
+		long[] assetTagsIds = ListUtil.toLongArray(
+			assetTags, AssetTag.TAG_ID_ACCESSOR);
 
-		if (group == null) {
-			return groupId;
-		}
-
-		return group.getGroupId();
+		document.addKeyword(Field.ASSET_TAG_IDS, assetTagsIds);
 	}
 
 	@Reference
-	protected GroupLocalService groupLocalService;
+	protected AssetTagLocalService assetTagLocalService;
+
+	@Reference
+	protected Portal portal;
 
 }
